requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14
djangorestframework-simplejwt[crypto]==5.3.1
drf-spectacular==0.27.1
psycopg[binary]==3.1.12
sqlparse==0.4.4